+++ .eslintrc.json
@@ -7,6 +7,7 @@
 	"rules": {
 		"capitalized-comments": 0,
 		"@next/next/no-img-element": 0,
+		"@typescript-eslint/no-magic-numbers": 0,
 		"import-helpers/order-imports": [
 			"warn",
 			{
