+++ package.json
@@ -10,6 +10,7 @@
 	},
 	"dependencies": {
 		"axios": "^0.27.2",
+		"eslint-plugin-n": "^15.2.5",
 		"eslint-plugin-node": "^11.1.0",
 		"eslint-plugin-promise": "^6.0.1",
 		"next": "12.2.5",
@@ -22,8 +23,8 @@
 	"devDependencies": {
 		"@babel/core": "^7.18.13",
 		"@techmmunity/eslint-config": "^5.2.4",
-		"@types/node": "18.7.13",
-		"@types/react": "18.0.17",
+		"@types/node": "18.7.14",
+		"@types/react": "18.0.18",
 		"@types/react-dom": "18.0.6",
 		"@types/styled-components": "^5.1.26",
 		"eslint": "8.22.0",
