+++ src/web/components/Button/types.ts
@@ -1,0 +1,6 @@
+import type { ButtonHTMLAttributes, ReactNode } from "react";
+
+export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+	schema: "primary" | "secondary";
+	children: ReactNode;
+}
