+++ src/web/components/Header/Header.index.tsx
@@ -1,3 +1,5 @@
+import Link from "next/link";
+
 import { AiOutlineArrowLeft } from "react-icons/ai";
 
 import * as S from "./Header.styled";
@@ -5,9 +7,11 @@ import * as S from "./Header.styled";
 export const Header = () => {
 	return (
 		<S.Container>
-			<S.BackButton type="button" aria-details="get back to home">
-				<AiOutlineArrowLeft />
-			</S.BackButton>
+			<Link href="/">
+				<S.BackButton type="button" aria-details="get back to home">
+					<AiOutlineArrowLeft />
+				</S.BackButton>
+			</Link>
 		</S.Container>
 	);
 };
