+++ src/web/components/SearchInput/index.tsx
@@ -1,0 +1,16 @@
+import type { FC, InputHTMLAttributes } from "react";
+import { BiSearch } from "react-icons/bi";
+
+import * as S from "./styled";
+
+export const SearchInput: FC<InputHTMLAttributes<HTMLInputElement>> = ({
+	type = "search",
+	...rest
+}) => {
+	return (
+		<S.Container>
+			<S.SearchInputStyle type={type} {...rest} />
+			<BiSearch />
+		</S.Container>
+	);
+};
