+++ src/web/pages/Country/TimeZoneCard/TimeZoneCard.styled.ts
@@ -1,0 +1,27 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+	gap: 1.2rem;
+
+	h5 {
+		font-weight: 300;
+		font-size: 1.6rem;
+		line-height: 2rem;
+		letter-spacing: 0.15rem;
+	}
+
+	div {
+		font-size: 4.2rem;
+	}
+
+	span {
+		font-weight: 300;
+		font-size: 2.2rem;
+		line-height: 2.6rem;
+		letter-spacing: 0.25rem;
+	}
+`;
