+++ src/web/pages/Country/TimeZoneCard/TimeZoneCard.types.ts
@@ -1,0 +1,7 @@
+export interface TimeZoneCardProps {
+	timeZoneType: "afternoon" | "dawn" | "morning" | "night";
+	predicatedTemperature: string;
+	predicatedStatusIconUrl: string;
+	predicatedStatusIconAlt: string;
+	temperatureVersion: "celsius" | "fahrenheit";
+}
