+++ src/web/pages/HomePage/HomePage.index.tsx
@@ -1,126 +1,51 @@
 /* eslint-disable @typescript-eslint/no-magic-numbers */
-import type { FC } from "react";
+import { useRouter } from "next/router";
+
+import type { FC, KeyboardEvent } from "react";
 import { useState } from "react";
-import { CgArrowsV } from "react-icons/cg";
 
-import { Header } from "web/components/Header/Header.index";
-
-import type { HomePageProps } from "./HomePage.types";
+import { Button } from "web/components/Button";
+import { SearchInput } from "web/components/SearchInput";
 
 import * as S from "./HomePage.styled";
-import { TimeZoneCard } from "./TimeZoneCard/TimeZoneCard.index";
 
-export const HomePage: FC<HomePageProps> = ({ data }) => {
-	const [temperatureVersion, setTemperatureVersion] = useState<
-		"celsius" | "fahrenheit"
-	>("celsius");
-	const [forecastDay] = data.forecast.forecastday;
+export const HomePage: FC = () => {
+	const [searchCountry, setSearchCountry] = useState("");
+	const router = useRouter();
 
-	const selectedDayHours = [3, 9, 15, 21];
-
-	const filteredTimePeriod = forecastDay.hour.filter((_, idx) => {
-		return selectedDayHours.some(selectedIndex => selectedIndex === idx);
-	});
-
-	const toggleTemperatureVersion = () => {
-		const newTemperature =
-			temperatureVersion === "celsius" ? "fahrenheit" : "celsius";
-
-		setTemperatureVersion(newTemperature);
+	const updateSearchCountry = (newValue: string) => {
+		setSearchCountry(newValue);
 	};
 
-	const getTimeZoneTypeText = (idx: number | string) => {
-		switch (idx) {
-			case 0:
-				return "dawn";
-			case 1:
-				return "morning";
-			case 2:
-				return "afternoon";
-			default:
-				return "night";
+	const HandleSearch = () => {
+		router.push(`/countries/${searchCountry}`);
+	};
+
+	const SearchOnEnter = (e: KeyboardEvent<HTMLButtonElement>) => {
+		if (e.key === "Enter") {
+			HandleSearch();
 		}
 	};
 
 	return (
-		<S.Container>
-			<Header />
-			<S.CityDetailsContainer onClick={toggleTemperatureVersion}>
-				<h2>{data.location.name}</h2>
-				<h4>{data.current.condition.text}</h4>
+		<S.Container onKeyDown={SearchOnEnter}>
+			<h1>ForecastThi</h1>
 
-				<S.TemperatureContainer role="button">
-					<h3>
-						{temperatureVersion === "celsius"
-							? data.current.temp_c
-							: data.current.temp_f}
-					</h3>
-					<span>
-						<CgArrowsV />
-					</span>
-					<div>
-						<span>{temperatureVersion === "celsius" ? "°C" : "°F"}</span>
-						<span>
-							{temperatureVersion === "celsius"
-								? forecastDay.day.maxtemp_c
-								: forecastDay.day.maxtemp_f}
-							°
-						</span>
-						<span>
-							{temperatureVersion === "celsius"
-								? forecastDay.day.mintemp_c
-								: forecastDay.day.mintemp_f}
-							°
-						</span>
-					</div>
-				</S.TemperatureContainer>
+			<p>
+				Search for your city using the US Zip code, UK Postcode, Canada Postal
+				code, IP address, Latitude/Longitude (decimal degree) or city name.
+			</p>
 
-				<S.StatusIconContainer>
-					<img
-						src={data.current.condition.icon}
-						alt={data.current.condition.text}
-					/>
-				</S.StatusIconContainer>
-				<S.predicatedTemperatureContainer>
-					{filteredTimePeriod.map((hourDetails, idx) => {
-						const timeZoneType = getTimeZoneTypeText(idx);
-
-						return (
-							<TimeZoneCard
-								key={hourDetails.time_epoch.toString()}
-								timeZoneType={timeZoneType}
-								predicatedTemperature={
-									temperatureVersion === "celsius"
-										? hourDetails.temp_c.toString()
-										: hourDetails.temp_f.toString()
-								}
-								predicatedStatusIconUrl={hourDetails.condition.icon}
-								predicatedStatusIconAlt={hourDetails.condition.text}
-								temperatureVersion={temperatureVersion}
-							/>
-						);
-					})}
-				</S.predicatedTemperatureContainer>
-
-				<S.WeatherTemperatureContainer>
-					<S.WeatherTemperatureItem>
-						<h5>wind speed</h5>
-						<span>{data.current.wind_mph} m/s</span>
-					</S.WeatherTemperatureItem>
-					<S.WeatherTemperatureItem>
-						<h5>sunrise</h5>
-						<span>{forecastDay.astro.sunrise}</span>
-					</S.WeatherTemperatureItem>
-					<S.WeatherTemperatureItem>
-						<h5>sunset</h5>
-						<span>{forecastDay.astro.sunset}</span>
-					</S.WeatherTemperatureItem>
-					<S.WeatherTemperatureItem>
-						<h5>humidity</h5>
-						<span>{data.current.humidity} %</span>
-					</S.WeatherTemperatureItem>
-				</S.WeatherTemperatureContainer>
-			</S.CityDetailsContainer>
+			<S.SearchContainer>
+				<SearchInput
+					placeholder="Search..."
+					onChange={e => updateSearchCountry(e.target.value)}
+					value={searchCountry}
+				/>
+				<Button onClick={HandleSearch} schema="primary" type="button">
+					Search
+				</Button>
+			</S.SearchContainer>
 		</S.Container>
 	);
 };
